pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "tetsuya-site"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-allauth = "^0.52.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"
dj-database-url = "^1.2.0"
dj-rest-auth = "^3.0.0"
psycopg2 = "2.9.3"
boto3 = "^1.26.79"
django-storages = "^1.13.2"
django-ses = "^3.3.0"


